requirements.txt
altgraph==0.17.4
APScheduler==3.11.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
idna==3.10
lockfile==0.12.2
packaging==25.0
pycparser==2.22
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
python-daemon==3.1.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
tzlocal==5.3.1
urllib3==2.5.0
